expand economy screen, split economy file a bit
