+++ src/gui/alice_ui.hpp
@@ -2,6 +2,7 @@
 #include "gui_element_base.hpp"
 #include "system_state.hpp"
 #include "economy_government.hpp"
+#include "economy_stats.hpp"
 
 namespace alice_ui {
 namespace budget_categories {
