+++ src/launcher/launcher_main.cpp
@@ -17,6 +17,7 @@
 #include "province.cpp"
 #include "triggers.cpp"
 #include "effects.cpp"
+#include "economy_stats.cpp"
 #include "economy.cpp"
 #include "economy_government.cpp"
 #include "demographics.cpp"
